+++ crates/wdl-engine/src/stdlib/quote.rs
@@ -32,7 +32,8 @@ fn quote(context: CallContext<'_>) -> Result<Value, Diagnostic> {
         .map(|v| match v {
             Value::None => PrimitiveValue::new_string("\"\"").into(),
             Value::Primitive(v) => {
-                PrimitiveValue::new_string(format!("\"{v}\"", v = v.raw())).into()
+                PrimitiveValue::new_string(format!("\"{v}\"", v = v.raw(Some(context.context))))
+                    .into()
             }
             _ => panic!("expected an array of primitive values"),
         })